requirements.txt
streamlit>=1.28.0
anthropic>=0.8.0
openai>=1.3.0
python-pptx>=0.6.21
python-docx>=0.8.11
openpyxl>=3.1.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
httpx>=0.25.0
Pillow>=10.0.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
pyyaml>=6.0
matplotlib>=3.7.0
flask>=2.3.0
flask-cors>=4.0.0
lxml>=4.9.0
